# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NARA Academic Schedule & WhatsApp Reminder API",
        "description": "REST API for managing semesters, courses, lecturers, PICs, schedules, and automated WhatsApp reminders via Fonnte.",
        "contact": {
            "email": "admin@nara.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "NARA API Server"
        }
    ],
    "paths": {
...
        "/api/semesters": {
            "get": {
                "tags": [
                    "Semesters"
                ],
                "summary": "Get list of semesters",
                "description": "Returns list of all semesters",
                "operationId": "getSemestersList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Semesters"
                ],
                "summary": "Create new semester",
                "description": "Creates a new semester record",
                "operationId": "storeSemester",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2026/2027 Ganjil"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Semester created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/semesters/{id}": {
            "get": {
                "tags": [
                    "Semesters"
                ],
                "summary": "Get semester by ID",
                "description": "Returns single semester data",
                "operationId": "getSemesterById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Semesters"
                ],
                "summary": "Update existing semester",
                "description": "Updates a semester record",
                "operationId": "updateSemester",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2026/2027 Genap"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Semester updated successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Semesters"
                ],
                "summary": "Delete semester",
                "description": "Deletes a semester record",
                "operationId": "deleteSemester",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Semester deleted successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
...
}